requirements.txt
langchain
sentence-transformers
faiss-cpu
python-dotenv
pypdf
python-docx
transformers
accelerate
langchain-community
langchain-huggingface
huggingface_hub[hf_xet]
huggingface_hub
gradio

